libgit: introduce deleted-repo cleaner
Run it after every git operation, rather than periodically, because it
would be difficult and expensive to build something that crawls
*every* TLF for a user to look for repos to delete.

When a repo is deleted, this schedules a background task to clean up
the repo after a set amount of time.  If KBFS is shut down before that
scheduled task, any other git operation in that folder will trigger a
cleanup.  However, if the cleanup blocks a git operation, it is
time-limited, and will only delete as many files as it can within a 2
second timeout, so as not to block foreground operation for too long.
Subsequent operations will clean up more data, etc, until the full
repo is gone.

Issue: KBFS-2442